+++ src/App/AppUI.js
@@ -1,33 +1,24 @@
-import React,{Fragment} from "react";
-
+import React, { Fragment, useContext } from "react";
+import { TodoContext } from "../TodoContext";
 import { CreateTodoButton } from "../CreateTodoButton";
 import { TodoCounter } from "../TodoCounter";
 import { TodoItem } from "../TodoItem";
 import { TodoList } from "../TodoList";
 import { TodoSearch } from "../TodoSearch";
 
-function AppUI({
-  loading,
-  error,
+function AppUI() {
+  const { error, loading, searchedTodos, completeTodo, deleteTodo } =
+    useContext(TodoContext);
 
-  totalTodos,
-  todosCompleted,
-  searchValue,
-  setSearchValue,
-  searchedTodos,
-  completeTodo,
-  deleteTodo,
-}) {
   return (
     <Fragment>
-      <TodoCounter totalTodos={totalTodos} todosCompleted={todosCompleted} />
+      <TodoCounter  />
 
-      <TodoSearch searchValue={searchValue} setSearchValue={setSearchValue} />
-
+      <TodoSearch />
       <TodoList>
         {error && <p>ERROR</p>}
         {loading && <p>Estamos cargando ,no desesperes</p>}
-        {(!loading  && !searchedTodos.lenght)&& <p>Crea tu primer todo</p>}   
+        {!loading && !searchedTodos.lenght && <p>Crea tu primer todo</p>}
         {searchedTodos.map((todo) => (
           <TodoItem
             key={todo.text}
@@ -38,7 +29,6 @@ function AppUI({
           />
         ))}
       </TodoList>
-
       <CreateTodoButton />
     </Fragment>
   );
