+++ src/TodoContext/index.js
@@ -1,0 +1,64 @@
+import { createContext,useState } from "react";
+import { useLocalStorage } from "../App/useLocalStorage";
+const TodoContext = createContext();
+
+function TodoProvider(props) {
+  const {
+    item: todos,
+    saveItem: saveTodos,
+    loading,
+    error,
+  } = useLocalStorage("TODOS_V1", []);
+
+  const [searchValue, setSearchValue] = useState("");
+  const todosCompleted = todos.filter((todo) => !!todo.completed).length;
+  const totalTodos = todos.length;
+
+  let searchedTodos = [];
+
+  if (!searchValue.length >= 1) {
+    searchedTodos = todos;
+  } else {
+    searchedTodos = todos.filter((todo) => {
+      const todoText = todo.text.toLowerCase();
+      const searchText = searchValue.toLowerCase();
+      return todoText.includes(searchText);
+    });
+  }
+
+  const completeTodo = (text) => {
+    const todoIndex = todos.findIndex((todo) => todo.text === text);
+    const newItem = [...todos];
+
+    newItem[todoIndex].completed = true;
+    saveTodos(newItem);
+  };
+
+  const deleteTodo = (text) => {
+    const todoIndex = todos.findIndex((todo) => todo.text === text);
+    const newItem = [...todos];
+
+    newItem.splice(todoIndex, 1);
+    saveTodos(newItem);
+  };
+
+  return (
+    <TodoContext.Provider
+      value={{
+        loading,
+        error,
+        totalTodos,
+        todosCompleted,
+        searchValue,
+        setSearchValue,
+        searchedTodos,
+        completeTodo,
+        deleteTodo,
+      }}
+    >
+      {props.children}
+    </TodoContext.Provider>
+  );
+}
+export {TodoContext,TodoProvider}
+
