+++ src/TodoCounter/index.js
@@ -1,6 +1,9 @@
-import React from "react";
+import React, { useContext } from "react";
 import './TodoCounter.css';
-function TodoCounter( {todosCompleted,totalTodos}) {
+import { TodoContext } from "../TodoContext";
+
+function TodoCounter( ) {
+    const {todosCompleted,totalTodos} = useContext(TodoContext);
     return(
         <h2 className="TodoCounter">Has completado {todosCompleted} de {totalTodos} TODOs</h2>
     );}export {TodoCounter}
