+++ src/TodoSearch/index.js
@@ -1,8 +1,9 @@
-import React from "react";
+import React, { useContext } from "react";
 import './TodoSearch.css';
+import { TodoContext } from "../TodoContext";
 
-function TodoSearch({searchValue, setSearchValue}) {
-
+function TodoSearch() {
+  const {searchValue, setSearchValue} = useContext(TodoContext);
   const onSearchValueChange=(event)=>{
    setSearchValue(event.target.value);
   }
